+++ app/controllers/grape_swagger_rails/application_controller.rb
@@ -1,12 +1,19 @@
 module GrapeSwaggerRails
   class ApplicationController < ActionController::Base
-    before_filter do
-      if GrapeSwaggerRails.options.before_filter
-        instance_exec(request, &GrapeSwaggerRails.options.before_filter)
-      end
+    if Rails::VERSION::MAJOR >= 4
+      before_action { run_before_action }
+    else
+      before_filter { run_before_action }
     end
 
     def index
+    end
+
+    private
+
+    def run_before_action
+      return unless GrapeSwaggerRails.options.before_action
+      instance_exec(request, &GrapeSwaggerRails.options.before_action)
     end
   end
 end