+++ spec/features/swagger_spec.rb
@@ -101,7 +101,18 @@ describe 'Swagger' do
     end
     context '#before_filter' do
       before do
-        GrapeSwaggerRails.options.before_filter do |_request|
+        allow(ActiveSupport::Deprecation).to receive(:warn)
+      end
+      it 'throws deprecation warning' do
+        GrapeSwaggerRails.options.before_filter { true }
+
+        expect(ActiveSupport::Deprecation).to have_received(:warn).with('This option is deprecated ' \
+          'and going to be removed in 0.3.0. Please use `before_action` instead')
+      end
+    end
+    context '#before_action' do
+      before do
+        GrapeSwaggerRails.options.before_action do |_request|
           flash[:error] = 'Unauthorized Access'
           redirect_to '/'
           false